+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ## v2.0.1 (Dec 13 2025)
 
-- **fix**: Air pressure device MA10238 had wrong humidity key (error) #40
-- **fix**: Fix unit in rain flip counter from mm to none. Count history. Needed a separate sensor. #41
-- **fix**: Add sensor MA10250PRO in device list with existing sensors types #40
+- **fix**: Air pressure device MA10238 had wrong humidity key (error) [#40](https://github.com/CestLaGalere/mobilealerts/issues/40)
+- **fix**: Fix unit in rain flip counter from mm to none. Count history. Needed a separate sensor. [#41](https://github.com/CestLaGalere/mobilealerts/issues/41)
+- **feat**: Add sensor MA10250PRO in device list with existing sensors types [#40](https://github.com/CestLaGalere/mobilealerts/issues/40)
+- **feat**: Add sensor MA10410 weather station [#45](https://github.com/CestLaGalere/mobilealerts/issues/45)
 - **chore**: Adapt readme.md with better log info and describe how to count the rain sensor (with use of Utility Meters)